Crate matroska

A library for Matroska file metadata parsing functionality

Implemented as a set of nested structs with public values which one can use directly.

Example

use std::fs::File;
use matroska::Matroska;
let f = File::open("filename.mkv").unwrap();
let matroska = Matroska::open(f).unwrap();
println!("title : {:?}", mkv.info.title);

For additional information about the Matroska format, see the official specification

Structs

Attachment
An attached file (often used for cover art)

Audio
An audio track's specifications

Chapter
An individual chapter point

ChapterDisplay
The display string for a chapter point entry

ChapterEdition
A complete set of chapters

Info
An Info segment with information pertaining to the entire file

Matroska
A Matroska file

Track
A TrackEntry segment in the Tracks segment container

Video
A video track's specifications

Enums

MatroskaError
A possible error when parsing a Matroska file

Settings
The settings a track may have

Tracktype
The type of a given track